requirements.txt
absl-py==2.1.0
aiofiles==23.2.1
albumentations==1.4.7
altair==5.3.0
annotated-types==0.6.0
anyio==4.2.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.4.1
async-lru==2.0.4
attrs==23.1.0
Babel==2.11.0
beautifulsoup4==4.12.2
bleach==4.1.0
Brotli==1.0.9
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==2.0.4
click==8.1.7
colorama==0.4.6
comm==0.2.2
contourpy==1.2.1
cycler==0.12.1
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
dnspython==2.6.1
email_validator==2.1.1
enum-compat==0.0.3
exceptiongroup==1.2.0
executing==2.0.1
fastapi==0.111.0
fastapi-cli==0.0.3
fastjsonschema==2.16.2
ffmpy==0.3.2
filelock==3.13.1
fonttools==4.52.1
fsspec==2024.5.0
gradio==4.31.5
gradio_client==0.16.4
grpcio==1.63.0
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.23.1
idna==3.4
imageio==2.34.1
importlib_metadata==7.1.0
importlib_resources==6.4.0
ipykernel==6.29.3
ipython==8.24.0
jedi==0.19.1
Jinja2==3.1.3
joblib==1.4.2
json5==0.9.6
jsonschema==4.19.2
jsonschema-specifications==2023.7.1
# jupyter_client==8.6.1
# jupyter_core==5.5.0
# jupyter-events==0.8.0
# jupyter-lsp==2.2.0
# jupyter_server==2.10.0
# jupyter_server_terminals==0.4.4
# jupyterlab==4.0.11
# jupyterlab-pygments==0.1.2
# jupyterlab_server==2.25.1
kiwisolver==1.4.5
lazy_loader==0.4
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.9.0
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==2.0.4
mkl-fft==1.3.8
mkl-random==1.2.4
mkl-service==2.4.0
mpmath==1.3.0
nbclient==0.8.0
nbconvert==7.10.0
nbformat==5.9.2
nest_asyncio==1.6.0
networkx==3.1
notebook_shim==0.2.3
numpy==1.26.4
opencv-python-headless==4.9.0.80
orjson==3.10.3
overrides==7.4.0
packaging==24.0
pandas==2.2.2
pandocfilters==1.5.0
parso==0.8.4
pickleshare==0.7.5
pillow==10.2.0
pip==23.3.1
platformdirs==4.2.1
prometheus-client==0.14.1
prompt-toolkit==3.0.42
protobuf==5.26.1
psutil==5.9.0
pure-eval==0.2.2
pycparser==2.21
pydantic==2.7.1
pydantic_core==2.18.2
pydub==0.25.1
Pygments==2.18.0
pyparsing==3.1.2
PySocks==1.7.1
# python-dateutil==2.9.0
# python-dotenv==1.0.1
python-json-logger==2.0.7
python-multipart==0.0.9
pytz==2024.1
# pywin32==305.1
# pywinpty==2.0.10
PyYAML==6.0.1
pyzmq==25.1.2
referencing==0.30.2
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.1
rpds-py==0.10.6
ruff==0.4.5
scikit-image==0.23.2
scikit-learn==1.4.2
scipy==1.13.0
semantic-version==2.10.0
Send2Trash==1.8.2
setuptools==68.2.2
shellingham==1.5.4
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
stack-data==0.6.2
starlette==0.37.2
sympy==1.12
tensorboard==2.16.2
tensorboard-data-server==0.7.2
terminado==0.17.1
threadpoolctl==3.5.0
tifffile==2024.5.10
tinycss2==1.2.1
tomlkit==0.12.0
toolz==0.12.1
torch==2.2.1
# torch-model-archiver==0.11.0
# torch-workflow-archiver==0.2.13
# torchaudio==2.2.1
# torchserve==0.11.0
torchsummary==1.5.1
torchvision==0.17.1
tornado==6.3.3
tqdm==4.66.4
traitlets==5.14.3
typer==0.12.3
typing_extensions==4.9.0
tzdata==2024.1
ujson==5.10.0
urllib3==2.1.0
uvicorn==0.29.0
watchfiles==0.21.0
wcwidth==0.2.13
webencodings==0.5.1
websocket-client==1.8.0
websockets==11.0.3
Werkzeug==3.0.3
wheel==0.41.2
win-inet-pton==1.1.0
zipp==3.17.0
